test(tui): scene footer detail goldens
